requirements.txt
Django==4.0.3
djangorestframework==3.13.1
django-cors-headers==3.12.0
dj-database-url==0.5.0
drf-nested-routers
Pillow==9.1.1
gunicorn==20.1.0
numpy==1.20.3
celery==5.2.7
whitenoise==6.1.0
redis
pymysql
flower
keras==2.7.0
tensorflow-cpu==2.7.0
eventlet==0.33.1
protobuf<=3.20.1
postgres
cython
postgis
psycopg2
djangorestframework-gis==1.0